No bug - Factor high-level interface to Preprocessor.py out of preprocess.py.
